Wine has a bug when Retina mode is enabled where the "window" is incorrectly sized, and the mouse cursor misaligned if launched in Full Screen/Borderless. The user normally has to switch from say, Borderless->Windowed->Borderless to fix it. But if they aren't a controller user this is made quite difficult due to the mouse cursor misalignment.
This change enables XOM to assist in the work around by changing the Screen Mode to windowed on every launch, so the user just has to re-select Borderless/Full Screen on the title screen. It also avoids the mouse misalignment. The user will only be asked once to enable this automatically, but the setting appears in preferences.